requirements.txt
Django>=4.2
psycopg2-binary
celery
django-celery-beat
djangorestframework
gunicorn
redis
django-cors-headers>=4.0.0
django-filter>=23.1
